requirements.txt
pandas~=2.2.2
dask~=2024.5.0
spacy~=3.7.2
faiss-cpu~=1.8.0
numpy~=1.26.4
scikit-learn~=1.5.1
sentence-transformers~=3.1.1
chromadb~=0.5.11